Update tests in database package
The database package reads in testdata which is prepended with the legacy
network magic. Since the tests were comparing it to the new magic the tests
were failing. For now we'll just make the tests use the old magic. In the
future we will probably want to update the test data to use the new magic.