Update tests in the blockchain package
The blockchain package tests read test data which is prepended with the old
network magic. Since we updated the default network magic this caused the tests
to fail. For now we will just use the old network magic but in the future we'll
probably want to build new test data with the new magic.