Add hardfork activations to params
This adds the hardfork block heights for the UAHF, and DAA hardforks
as well as sets the activation time for MagneticAnomaly and GreatWall.

Nothing in the code uses these parameters at the moment but will going forward
as we implement the Bitcoin Cash forks.

Note the May 15, 2018 hardfork ("monolith") retroactively activated opcodes
going all the way back to genesis. So technically we do not need any code
regarding activation though we don't yet have the code for the opcodes.