Add excessiveblocksize as a config option
This brings bchd in line with the other implementations in that it allows the user
to configure the max block size limit without needing developer approval.

We do not allow excessive block sizes less than 32MB as that is the current
network consensus, however the user may set the value larger. Setting the value
larger is discourage right now however, as most other users and miners are not
running with an EB larger than 32MB.