Make blocksize and sigops limits configurable
This commit implements the blocksize and sigops consensus rule changes for the
August 1st, 2017 (Uahf) hardfork. The max block size limit is added to the conif
as ExcessiveBlockSize and the various consensus rule checks are made to use it.

According to the rules, the 1mb legacy blocksize limit must still be enforced
prior to the hardfork and must change to the EB limit after the fork.