Fix blockchain package tests
This updates the blockchain package tests to incorporate the blocksize changes.
Currently the tests are using 1MB as the max blocksize. This is OK for now as
the blocksize is technically configurable, but we should make the tests work
with large mock blocksizes. The primary issue with doing this is creating a
32MB block. Currently the tests only generate a single transaction per block
and since the transaction cannot be larger than 1MB you cannot make a block
larger than 1MB. The tests would need to be refactored to add more transactions
per block but this is very tedious to do because the tests are all set up such
that the transactions are all valid spends of each other and simply creating
more transactions means basically refactoring the entire chain of transactions.

Also this commit move the blocksize check into the checkBlockContext function
as that function is intended for checks that depend on the blocks position in
the chain.