Fix block sorting alorithm
There was a bug in calculating the suitable block for difficulty adjustments that
would arise when two blocks had the same timestamp. Apparently there aren't any
on mainnet so this bug only happened on testent. It doesn't look like the sort
function in the standard library is capable of handling this correctly so
we will just need to sort the blocks the same way ABC does it.