Prevent sync peer disconnecting during a cache flush
Utxo cache flushes can take a few minutes to complete. The code currently disconnects
the sync peer if there's not activity for a period of time. This commit exposes
the state of the cache flushing from the Blockchain package so that we can
avoid disconnecting the sync peer if we're in the middle of a flush.