Rollback blocks anytime utxo state is inconsistent
Ideally we would only rollback the state when we had a hard shutdown in the middle
of a flush but since the underlying db also has a cache that might not write
our flushing state if there's a hard shutdown, we might not be able to detect
that we shutdown in the middle of a flush. So to make sure we always restore
properly we will always rollback the blockchain when the state is inconsistent.