Respond to GetBlocks with NotFound when in pruned mode
Before this commit if the node recieved a GetBlocks message it would respond
with an inv of the needed blocks. However, in prune mode we might not have
those blocks so we don't want to mislead the remote peer. Typically nodes
do not advertise NodeNetwork when in prune mode so they wont recieve GetBlocks
requests. However, we intend to implement NodeNetworkLimited which means we
may receive GetBlock requests and we don't want to force the remote peer to
timeout before moving on to try to get blocks from someone else. So when in
prune mode we'll respond with NotFound if the requested blocks are below our
prune depth.