Add force config option for deep rollbacks
Calculating the utxo hash of a deep block may use a lot of memory since the
utxos for each block that gets rolled back are held in memory. To prevent
unwanted memory exhaustion we will return an error if they are attempting to
rollback deeper than the last checkpoint. If they wish to do this anyway a new
--force option is created that will permit this.