Fix utxo reorg bug
We hit a bug with the following setup. Block 559307(a) was mined creating a new
utxo (u). An alternate block 559307(b) was mined which did not create u. Subsequently
559308 was mined on top of b which did create u again. So during the reorg u
needed to go from unspent to spent(removed) when block a was disconnected, then
back to unspent again when block 559308 was connected. However, the utxocache
commit() function did not have the ability to override a utxo previously marked
as spent back into an unspent state and thus u was left marked as spent. When
559309 was mined spending u, it failed to validate because u was erroneously marked
as spent already.

This commit patches the commit() function to allow overriding from spent to
unspent.