Save blockchain type to database and use it to set service flags
This commit adds a new db entry to the metadata bucket to save the type of
blockchain (pruned, full, etc). If the chain was ever run in pruned mode or
fastsync mode then we need to set the service flag to NodeNetworkLimited. A
new BlockChain method is added for this purpose.