Update GetCFMempool previous script calculation
The function as it was written only checked the utxo cache for scripts and
didn't check the mempool itself. This means it would possibly omit prevout scripts
from transactions spending from other transactions in the mempool.

This commit fixes the issue by creating new exported mempool method to get
both the txs and scripts.