Refactor ExtractDataElements to be more explict
ExtractDataElements only extracts elements for data carrying opcodes. Some
opcodes such as OP_0 or OP_16 also push data to the stack but our implementation
does not treat them as data carrying opcodes because the "data" field in the
opcode struct is null for those those opcodes.

This function was evaluating the non-data carrying opcodes before, but just not
appending any bytes to the return since the "data" field was nil. This commit
refactors the function to only evaluate the opcodes which carry data.

The functionality remains the same but it's now more clear that our intention is
to only return data for data carrying opcodes.

The comment has been updated to reflect this.