Create indexer migrations
This commit creates a Migrate method in the indexer interface. It also adds
a migration to the cfilter index which just drops the index and rebuilds. The
purpose it make sure all nodes are using the same index as the filtering code
has changed recently and it appears some nodes have a different index.