Update to latest May hardfork spec
The May hardfork spec has changed to reflect a different implementation in
the bitcoin-abc codebase. It was determined it would be easier to get bchd
nodes to upgrade with a different spec than to get all bitcoin-abc nodes
to upgrade to the existing spec.

The specific change allows the top stack item to be zero when the script is
segwit exempt.