Override utxo entry in reorg case
The code here wants to override the existing entry in the cache in the reorg case.
However, the override bool was left to false so it wouldn't actually do the
override. Also, by setting ourEntry to entry rather than a copy of entry any modifications
to entry performed later (which happens at least when the view is pruned) will
modifiy the cache. This commit fixes both issues.