This commit updates the reference tests to use the exact same test vectors from Bitcoin-ABC.
https://github.com/Bitcoin-ABC/bitcoin-abc/blob/master/src/test/data/script_tests.json

Well, almost the exact same. The only two exceptions are:
1. We still use the ScriptVerifyCheckDataSig flag whereas ABC removed it in
favor of retroactive activation. Thus we must include the flag in all tests
using CHECKDATASIG or CHECKDATASIGVERIFY.
2. The two tests cases testing the automatic replay protection have been removed
since this feature is not implemented in bchd.

Other than this the file is identical. The test uncovered a bug in OP_CHECKMULTISIG
and OP_CHECKMULTISIGVERIFY where 64 byte sigantures were not being rejected.

This commit also fixes this bug.

It also implements ScriptVerifyMinimalIf and ScriptVerifyCompressedPubkey for
the sake of the tests.