Set indexManger to nil if chain is pruned
The fast sync feature is not playing nice with the recent migration due to
the chain being pruned. If someone runs their node without the fastsync flag
after previously doing so, it will get into a bad state.

This commit sets the indexManger to nil inside the blockchain package if it
was ever run in fast sync mode previously.