remove this TODO note
Doing what is described in this TODO note is
not necessary, and it would only add code complexity.

Since we have implemented Burn Flags in the grpc server,
there is an indication that there is a missing vout. So
developers can use that information to detected when mint
baton has been destroyed in the rare missing vout situation.