Peer/sync manager improvements (#470)
* [SyncManager] fix regtest isSyncCandidate check

Some node software does not advertise full node service
when in regtest mode.  We only want to check services
when we're not running in regression test mode.

* [SyncManager] recheck isSyncCandidate in startSync()

There was a problem with the previous code used to select a
sync peer when none exists, which happens if we lose our sync
peer or if we fail to establish a sync peer.

The early exit statement didn't first re-check if the peer is a
possible syncCandidate before exiting.  This would have been
fine if 'syncCandidate' wasn't also possibly being changed in
the following 'peer.LastBlock() < best.Height' conditional.

This fix allows peer nodes to become sync candidates at a later
time if they were ever marked as a non-sync peer candidate.  This
could happen if previously checked peer was behind, and then later
surpassed our node.